startup_insights/requirements.txt
streamlit
firecrawl-py
agno
google-genai
python-dotenv